Merge payment into order
This is a semplification that is justified by our use case:
every time a user checks-out its cart a new order is created. Thus each
order can only have one associated payment.
A consequence of such approach is that there could potentially be many
pending orders in the database having the same information - in case the
user doesn't complete the payment and go back to the cart. This can be
addressed by cleaning the entries that are older than a certain
threshold TBD.